voicelines changes

+++ LifeUnderPressure/Assets/_Scripts/UI/Voicelines/Voiceline.cs
@@ -8,7 +8,7 @@ public class Voiceline : ScriptableObject
     [TextArea(5, 5)]
     public string[] voicelines;
 
-    public int charPerSecond = 10;
+    public float charPerSecond = 10;
 
     [SerializeField] private EventReference[] OnVoicelineStart;
